backend/requirements.txt
uvicorn
fastapi
psycopg2-binary
sqlalchemy
asyncio
aiohttp
python-dotenv
Pillow
python-multipart
PyJWT
asyncpg